+++ executables/LUDUMDARE/Ludum45/src/Ludum45Particles.cpp
@@ -561,8 +561,8 @@ bool ParticleExplosionTrait::UpdateParticle(float delta_time, ParticleExplosion 
 	if (particle->explosion_info == nullptr) // delete the particle
 		return true;
 
-	int image_count = particle->explosion_info->GetAnimationImageCount();
-	float frame_time = particle->explosion_info->GetFrameTime();
+	size_t image_count = particle->explosion_info->GetAnimationImageCount();
+	float  frame_time = particle->explosion_info->GetFrameTime();
 
 	if (frame_time == 0)
 		frame_time = 1.0f / 16.0f;
