+++ libraries/death/include/death/GameInstance.h
@@ -42,9 +42,9 @@ namespace death
 		GameLevelInstance const * GetLevelInstance() const;
 
 		/** get the player by its index */
-		Player * GetPlayer(int player_index);
+		Player * GetPlayer(size_t player_index);
 		/** get the player by its index */
-		Player const * GetPlayer(int player_index) const;
+		Player const * GetPlayer(size_t player_index) const;
 
 		/** get the number of players */
 		size_t GetPlayerCount() const { return players.size(); }
